labelbox_backend/requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
certifi==2024.12.14
cffi==1.17.1
cloudinary==1.41.0
contextlib2==21.6.0
cryptography==44.0.0
Django==5.1.4
django-cors-headers==4.6.0
django-ninja==1.3.0
django-ninja-extra==0.21.8
django-ninja-jwt==5.3.5
dnspython==2.7.0
email_validator==2.2.0
idna==3.10
injector==0.22.0
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
PyJWT==2.10.1
python-decouple==3.8
ruff==0.8.4
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.3.0
whitenoise==6.8.2
